# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ISP Management Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /packages:
    get:
      operationId: listPackages
      tags: [packages]
      summary: List all internet packages
      responses:
        "200":
          description: List of packages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Package"
    post:
      operationId: createPackage
      tags: [packages]
      summary: Admin creates a new package
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PackageInput"
      responses:
        "201":
          description: Created package
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Package"
...
  /packages/{id}:
    patch:
      operationId: updatePackage
      tags: [packages]
      summary: Admin updates a package
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PackageUpdate"
      responses:
        "200":
          description: Updated package
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Package"
    delete:
      operationId: deletePackage
      tags: [packages]
      summary: Admin deletes a package
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── CUSTOMERS ─────────────────────────────────────────────────────────────
...
  /subscriptions:
    get:
      operationId: listSubscriptions
      tags: [subscriptions]
      summary: List subscriptions (admin sees all, customer sees own)
      parameters:
        - name: customerId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of subscriptions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subscription"
    post:
      operationId: createSubscription
      tags: [subscriptions]
      summary: Customer subscribes to a package
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriptionInput"
      responses:
        "201":
          description: Created subscription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
  /subscriptions/{id}:
    get:
      operationId: getSubscription
      tags: [subscriptions]
      summary: Get subscription details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Subscription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
    patch:
      operationId: updateSubscription
      tags: [subscriptions]
      summary: Admin updates subscription status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriptionUpdate"
      responses:
        "200":
          description: Updated subscription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
  /subscriptions/{id}/renew:
    post:
      operationId: renewSubscription
      tags: [subscriptions]
      summary: Customer requests renewal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Renewal subscription created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
  /subscriptions/{id}/switch:
    post:
      operationId: switchPackage
      tags: [subscriptions]
      summary: Customer requests package switch
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SwitchPackageInput"
      responses:
        "201":
          description: New subscription with new package
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"

  # ── PAYMENTS ──────────────────────────────────────────────────────────────
  /payments:
    get:
      operationId: listPayments
      tags: [payments]
      summary: List payments (admin sees all, customer sees own)
      parameters:
        - name: customerId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: subscriptionId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
    post:
      operationId: submitPayment
      tags: [payments]
      summary: Customer submits payment proof
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "201":
          description: Payment submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/{id}/verify:
    post:
      operationId: verifyPayment
      tags: [payments]
      summary: Admin verifies or rejects a payment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyPaymentInput"
      responses:
        "200":
          description: Payment updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"

  # ── COMPLAINTS ────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Package:
      type: object
      required: [id, name, speedMbps, price, validity, isActive, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        speedMbps:
          type: integer
        price:
          type: number
        validity:
          type: string
          enum: [monthly, quarterly, yearly]
        isActive:
          type: boolean
        description:
          type: ["string", "null"]
        createdAt:
          type: string
    PackageInput:
      type: object
      required: [name, speedMbps, price, validity]
      properties:
        name:
          type: string
        speedMbps:
          type: integer
        price:
          type: number
        validity:
          type: string
          enum: [monthly, quarterly, yearly]
        description:
          type: string
        isActive:
          type: boolean
    PackageUpdate:
      type: object
      properties:
        name:
          type: string
        speedMbps:
          type: integer
        price:
          type: number
        validity:
          type: string
          enum: [monthly, quarterly, yearly]
        description:
          type: string
        isActive:
          type: boolean
...
    Subscription:
      type: object
      required: [id, customerId, packageId, status, createdAt]
      properties:
        id:
          type: integer
        customerId:
          type: integer
        packageId:
          type: integer
        status:
          type: string
          enum: [pending-payment, active, suspended, expired]
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        createdAt:
          type: string
        package:
          $ref: "#/components/schemas/Package"
        customerName:
          type: ["string", "null"]
        customerPhone:
          type: ["string", "null"]
    SubscriptionInput:
      type: object
      required: [packageId]
      properties:
        packageId:
          type: integer
    SubscriptionUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending-payment, active, suspended, expired]
        startDate:
          type: string
        endDate:
          type: string
    SwitchPackageInput:
      type: object
      required: [newPackageId]
      properties:
        newPackageId:
          type: integer
    Payment:
      type: object
      required: [id, subscriptionId, customerId, amount, status, createdAt]
      properties:
        id:
          type: integer
        subscriptionId:
          type: integer
        customerId:
          type: integer
        amount:
          type: number
        proofImageUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, verified, rejected]
        adminNote:
          type: ["string", "null"]
        verifiedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        customerName:
          type: ["string", "null"]
        customerPhone:
          type: ["string", "null"]
        packageName:
          type: ["string", "null"]
    PaymentInput:
      type: object
      required: [subscriptionId, amount]
      properties:
        subscriptionId:
          type: integer
        amount:
          type: number
        proofImageUrl:
          type: string
        note:
          type: string
    VerifyPaymentInput:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [verify, reject]
        adminNote:
          type: string
...